restructure hudi-spark tests
